+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.3.3.0
+
+Version 0.3.3 now includes ways to directly interact and manage metadata programmatically.
+
+Also includes new classes to keep track of the analysis process.
+
 # 0.3.2.13
 
 Deal with numpy data types for when user wants to use datatype properties to connect from instances to literals